refactor: TripHeader -> TripPlanHeader로 이름 변경
