+++ src/components/trip/TripPlanHeader.vue
